% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDatos.R
\name{getDatos}
\alias{getDatos}
\title{Descarga todos las muestras}
\usage{
getDatos()
}
\value{
Un objeto de clase data.frame que incluye las lecturas de la base de datos.
}
\description{
Descarga todas las muestras de la base de datos de los programas de mejoramiento genetico de INIA.
}
\details{
Esta funcion corresponde solo a una prueba de concepto sobre el paso de formato de estructura de columna unica al formato tabular.
}
\examples{
library(inia.data)
# Descarga de datos
d <- getDatos()
class(d)
head(d)
}
\seealso{
La funcion \code{\link{getEnsayo}} muestra un ejemplo de comunicacion con el servicio web y \code{\link{getUbicacion}} muestra ejemplo de postproceso.
}
