% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getEnsayo.R
\name{getEnsayo}
\alias{getEnsayo}
\title{Descarga metadatos de un ensayo}
\usage{
getEnsayo(idEnsayo)
}
\arguments{
\item{idEnsayo}{Identificador del ensayo.}
}
\value{
Un objeto de clase data.frame con los metadatos del ensayo solicitado.
}
\description{
Descarga metadatos de un ensayo seleccionado de la base de datos de los programas de mejoramiento genetico.
}
\details{
Esta funcion es una prueba de concepto de una consulta a la base de datos con un cierto predicado: Devolver solo los valores correspondientes a una estacion.
}
\examples{
library(inia.data)
# Descarga de metadatos del ensayo
A <- getEnsayo(idEnsayo = 3)
B <- getEnsayo(idEnsayo = 4)
A
B
}
\seealso{
\code{\link{getUbicacion}}, \code{\link{getDatos}}
}
