% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getUbicacion.R
\name{getUbicacion}
\alias{getUbicacion}
\title{Devuelve un objeto que representa la ubicacion del ensayo}
\usage{
getUbicacion(idEnsayo)
}
\arguments{
\item{idEnsayo}{Identificador del ensayo.}
}
\value{
Un objeto con la posicion del ensayo de la clase \code{ubicacion}
}
\description{
Descarga la ubicacion de un ensayo
}
\details{
Esta funcion corresponde solo a una prueba de concepto sobre el paso de formato de estructura de columna unica al formato tabular.
}
\examples{
library(inia.data)
u <- getUbicacion(idEnsayo = 1)
cat("El centroide del ensayo se encuentra en ", u$centroid)
}
\seealso{
\code{\link{getEnsayo}}, \code{\link{getDatos}}
}
