% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.ubicacion.R
\name{plot}
\alias{plot}
\alias{plot.ubicacion}
\title{plot ubicacion object}
\usage{
\method{plot}{ubicacion}(x, ...)
}
\arguments{
\item{x}{Ubicacion del ensayo.}

\item{...}{Ignorado}
}
\description{
plot one or more attributes of an ubicacion object on a map
Plot ubicacion object
}
